API Reference

Garamond.ERRORED_REQUEST

Garamond.BruteTreeIndex

Garamond.HNSWIndex

Garamond.KDTreeIndex

Garamond.NaiveIndex

Garamond.SearchConfig

Garamond.build_corpus

Garamond.build_logger

Garamond.build_searcher

Garamond.construct_request

Garamond.construct_response

Garamond.deconstruct_request

Garamond.detect_language

Garamond.garamond_log_formatter

Garamond.get_parsing_function

Garamond.load_search_configs

Garamond.load_searchers

Garamond.meta2sv

Garamond.prepare_query

Garamond.printable_version

Garamond.recursive_glob

Garamond.rest_server

Garamond.search

Garamond.search

Garamond.search

Garamond.search_heuristically!

Garamond.search_server

Garamond.squash

Garamond.squash

Garamond.summarize

Garamond.unix_socket_server

Garamond.updater

Garamond.version

Garamond.web_socket_server

StringAnalysis.embed_document

StringAnalysis.embed_document

Garamond.SearchConfig — Type.

The search engine configuration object SearchConfig is used in building search objects of type Searcher and to provide information about them to other methods.

source

Garamond.load_searchers — Method.

load_searchers(configs)

Loads/builds searchers using the information provided by configs. The latter can be either a path, a vector of paths to searcher configuration file(s) or a vector of SearchConfig objects.

Loading process flow:

Parse configuration files using load_search_configs (if configs contains paths to configuration files)

The resulting Vector{SearchConfig} is passed to load_searchers (each SearchConfig contains the data filepath, parameters etc.)

Each searcher is build using build_searcher and a vector of searcher objects is returned.

source

Garamond.rest_server — Method.

rest_server(port::Integer, channel::Channel{String})

Starts a bi-directional REST server that uses the HTTP port port and communicates with the search server through a channel channel.

Service GET link format: /api/v1/<op>/<max_matches>/<search_method>/<max_suggestions>/<what_to_return>/<query> where: <op> is fixed to search <maxmatches> is a number larger than 0 <searchmethod> is fixed to exact <maxsuggestions> is fixed to 0 <whatto_return> is fixed to json-index <query> can be any string

Example: http://localhost:port/api/v1/search/100/exact/0/json-index/something%20to%20search

source

Garamond.search — Method.

search(index, point, k)

Searches for the k nearest neighbors of point in data contained in the index. The index may vary from a simple wrapper inside a matrix to more complex structures such as k-d trees, etc.

source

Garamond.search — Method.

search(srcher, query [;kwargs])

Searches for query (i.e. key terms) in a corpus' metadata, text or both and returns information regarding the the documents that match best the query. The function returns an object of type SearchResult and the id of the searcher.

Arguments

srcher::Searcher is the corpus searcher

query the query, can be either a String or Vector{String}

Keyword arguments

search_method::Symbol controls the type of matching: :exact searches for the very same string while :regex searches for a string in the corpus that includes the needle

max_matches::Int is the maximum number of search results to return

max_suggestions::Int is the maximum number of suggestions to return for each missing needle

source

Garamond.search — Method.

search(srcher, query [;kwargs])

Searches for query (i.e. key terms) in multiple corpora and returns information regarding the documents that match best the query. The function returns the search results in the form of a Vector{SearchResult}.

Arguments

srcher::Vector{Searcher} is the corpora searcher

query the query, can be either a String or Vector{String}

Keyword arguments

search_method::Symbol controls the type of matching: :exact searches for the very same string while :regex searches for a string in the corpus that includes the needle

max_matches::Int is the maximum number of search results to return from each corpus

max_suggestions::Int is the maximum number of suggestions to return for each missing needle from the search in a corpus

source

Garamond.unix_socket_server — Method.

unix_socket_server(socket::AbstractString, channel::Channel{String})

Starts a bi-directional unix socket server that uses a UNIX-socket socket and communicates with the search server through a channel channel.

source

Garamond.web_socket_server — Method.

web_socket_server(port::UInt16, channel::Channel{String})

Starts a bi-directional web socket server that uses a WEB-socket at port port and communicates with the search server through a channel channel.

source

Garamond.ERRORED_REQUEST — Constant.

Standard deconstructed request corresponding to an error request.

source

Garamond.BruteTreeIndex — Type.

BruteTree index type for storing text embeddings. It is a wrapper around a BruteTree NN structure and performs brute search using a distance-based similarity between vectors.

source

Garamond.HNSWIndex — Type.

HNSW index type for storing text embeddings. It is a wrapper around a HierarchicalNSW (Hierarchical Navigable Small Worlds [1]) NN graph structure and performs a very efficient search using a distance-based similarity between vectors. [1] Yu. A. Malkov, D.A. Yashunin "Efficient and robust approximate nearest neighbor search using Hierarchical Navigable Small World graphs" (https://arxiv.org/abs/1603.09320)

source

Garamond.KDTreeIndex — Type.

K-D Tree index type for storing text embeddings. It is a wrapper around a KDTree NN structure and performs a more efficient search using a distance-based similarity between vectors.

source

Garamond.NaiveIndex — Type.

Naive index type for storing text embeddings. It is a wrapper around a matrix of embeddings and performs brute search using the cosine similarity between vectors.

source

Garamond.build_corpus — Method.

build_corpus(documents, metadata_vector, doctype)

Builds a corpus of documents of type doctype using the data in documents and metadata from metadata_vector.

source

Garamond.build_logger — Function.

build_logger(logging_stream, log_level)

Builds a logger using the stream and loglevel provided. These should be coming from parsing the input arguments. The loggingstream can take the values: • "null": logs to /dev/null • "stdout": logs to standard output • "/path/to/existing/file": logs to an existing file • "/path/to/non-existing/file": creates the log file If no valid option is provided, the default stream is the standard output. The log level can take the values: "debug", "info", "error" and defaults to "info" if no valid option is provided.

source

Garamond.build_searcher — Method.

build_searcher(sconf::SearchConfig)

Creates a Searcher from a searcher configuration sconf.

source

Garamond.construct_request — Method.

construct_request(req::HTTP.Request)

Constructs a Garamond JSON search request from a HTTP request req: extracts the link, parses it, builds a dictionary and transforms it to JSON.

source

Garamond.construct_response — Method.

construct_response(srchers, results, what [; kwargs...])

Function that constructs a response for a Garamond client using the search results, data from srchers and specifier what.

source

Garamond.deconstruct_request — Method.

deconstruct_request(request)

Function that deconstructs a Garamond request received from a client into individual search engine operations and search parameters.

source

Garamond.detect_language — Method.

detect_language(text [; default=DEFAULT_LANGUAGE])

Detects the language of a piece of text. Returns a language of type Languages.Language. If the text is empty of the confidence is low, return the default language.

source

Garamond.garamond_log_formatter — Method.

garamond_log_formatter(level, _module, group, id, file, line)

Garamond -specific log message formatter. Takes a fixed set of input arguments and returns the color, prefix and suffix for the log message.

source

Garamond.get_parsing_function — Method.

get_parsing_function(args...)

Function that generates a parsing function from its input arguments and returns it.

Arguments

parser::Symbol is the name of the parser

parser_config::Union{Nothing, Dict} can contain optional configuration data for the parser (for delimited parsers)

header::Bool whether the file has a header or not (for delimited files only)

delimiter::String the delimiting character (for delimited files only)

globbing_pattern::String globbing pattern for gathering file lists from directories (for directory parsers only)

language::String the plain English name of the language; use "auto" for

document-level language autodetection

build_summary::Bool whether to use a summary instead of the full document (for directory parsers only)

summary_ns::Int how many sentences to use in the summary (for directory parsers only)

summarization_strip_flags::UInt32 flags used to strip text before summarization (for directory parsers only)

show_progress::Bool whether to show the progress when loading files

Note: parser must be in the keys of the PARSER_CONFIGS constant. The name of the data parsing function is created as: :__parser_<parser> so, the function name :__parser_delimited_format_1 corresponds to the parser :delimited_format_1. The function must be defined apriori.

source

Garamond.load_search_configs — Method.

load_search_configs(filename)

Creates search configuration objects from a data configuration file specified by filename. The file name can be either an AbstractString with the path to the configuration file or a Vector{AbstractString} specifying multiple configuration file paths. The function returns a Vector{SearchConfig} that is used to build the Searcher objects.

source

Garamond.meta2sv — Method.

meta2sv(metadata, fields=DEFAULT_METADATA_FIELDS)

Turns the metadata::DocumentMetadata object's fields into a vector of strings, where the value of each field becomes an element in the resulting vector.

source

Garamond.prepare_query — Method.

prepare_query(query, flags)

Prepares the query for search (tokenization if the case), pre-processing.

source

Garamond.printable_version — Method.

printable_version()

Returns a pretty version string that includes the git commit and date.

source

Garamond.recursive_glob — Function.

recursive_glob(pattern, path)

Globs recursively all the files matching the pattern, at the given path.

source

Garamond.search_heuristically! — Method.

search_heuristically!(suggestions, search_tree, needles [;max_suggestions=1])

Searches in the search tree for partial matches of the needles.

source

Garamond.search_server — Method.

search_server(data_config_paths, io_channel)

Search server for Garamond. It is a finite-state-machine that when called, creates the searchers i.e. search objects using the data_config_paths and the proceeds to looping continuously in order to: • update the searchers regularly; • receive requests from clients on the I/O channel • call search and route responses back to the clients through the I/O channel

Both searcher update and I/O communication are performed asynchronously.

source

Garamond.squash — Method.

squash(m)

Function that creates a single mean vector from a matrix m and performs some normalization operations as well.

source

Garamond.squash — Method.

squash(vv, m)

Function that creates a single mean vector from a vector of vectors vv where each vector has a length m and performs some normalization operations as well.

source

Garamond.summarize — Method.

summarize(sentences [;ns=1, flags=DEFAULT_SUMMARIZATION_STRIP_FLAGS])

Build a summary of the text's sentences. The resulting summary will be a ns sentence document; each sentence is pre-procesed using the flags option.

source

Garamond.updater — Method.

updater(searchers, channel, update_interval)

Function that regularly updates the searchers at each update_interval seconds, and puts the updates on the channel to be sent to the search server.

source

Garamond.version — Method.

version()

Returns the current Garamond version using the Project.toml and git. If the Project.toml, git are not available, the version defaults to an empty string.

source

StringAnalysis.embed_document — Method.

Function that embeds a document i.e. returns an embedding matrix, columns are word embeddings, using the Word2Vec or Glowe WordVectors object. The function has an identical signature as the one from the ConceptnetNumberbatch package.

source

StringAnalysis.embed_document — Method.

embed_document(embedder, lexicon, document [;
               embedding_method=DEFAULT_DOC2VEC_METHOD,
               isregex=false,
               sif_alpha=DEFAULT_SIF_ALPHA])

Function to get from multiple sentences to a document embedding. The embedding_method option controls how multiple sentence embeddings are combined into a single document embedding. Avalilable options for embedding_method: :bow - calculates document embedding as the mean of the sentence embeddings :sif - smooth-inverse-frequency subtracts paragraph/phrase vector from each sentence embedding

source