API Reference

Garamond.DTVModel

Garamond.ENVOP_REQUEST

Garamond.ERRORED_REQUEST

Garamond.EmbeddingsLibrary

Garamond.KILL_REQUEST

Garamond.READCONFIGS_REQUEST

Garamond.RESPONSE_TERMINATOR

Garamond.UNINITIALIZED_REQUEST

Garamond.BOEEmbedder

Garamond.BOREPEmbedder

Garamond.BruteTreeIndex

Garamond.CPMeanEmbedder

Garamond.DTVEmbedder

Garamond.DisCEmbedder

Garamond.HNSWIndex

Garamond.InternalRequest

Garamond.KDTreeIndex

Garamond.NaiveIndex

Garamond.SIFEmbedder

Garamond.SearchConfig

Garamond.SearchEnv

Garamond.SearchResult

Garamond.Searcher

Garamond.StringId

Base.length

Base.parse

Garamond.__found_missing_needles

Garamond.aggregate!

Garamond.build_data_env

Garamond.build_logger

Garamond.build_response

Garamond.build_result_from_ids

Garamond.build_search_env

Garamond.build_search_env

Garamond.build_searcher

Garamond.chop_to_length

Garamond.detect_language

Garamond.document2vec

Garamond.document2vec

Garamond.env_operator

Garamond.garamond_log_formatter

Garamond.noop_ranker

Garamond.parse_configuration

Garamond.printable_version

Garamond.query_preparation

Garamond.read_configuration_to_json

Garamond.respond

Garamond.rest_server

Garamond.search

Garamond.search

Garamond.search_server

Garamond.sentences2vec

Garamond.squash

Garamond.squash

Garamond.suggestion_search!

Garamond.summarize

Garamond.unix_socket_server

Garamond.version

Garamond.web_socket_server

Garamond.word_embeddings

HNSW.knn_search

Garamond.SearchConfig — Type.

The search engine configuration object SearchConfig is used in building search objects of type Searcher and to provide information about them to other methods.

source

Garamond.SearchResult — Type.

Object that stores the search results from a single searcher.

source

Garamond.Searcher — Type.

Search object. It contains all the indexed data and related

configuration that allows for searches to be performed.

source

Garamond.build_search_env — Method.

build_search_env(config_path; cache_path=nothing)

Creates a search environment using the information provided by the configuration file config_path.

source

Garamond.build_search_env — Method.

build_search_env(env_config; cache_path=nothing)

Creates a search environment using the information provided by the environment configuration env_config. A cache filepath can be specified by cache_path in which case the function will attempt to load it first.

source

Garamond.parse_configuration — Method.

parse_configuration(filename)

Creates search configuration objects from a data configuration file specified by filename. The file name can be either an AbstractString with the path to the configuration file or a Vector{AbstractString} specifying multiple configuration file paths. The function returns a Vector{SearchConfig} that is used to build the Searcher objects.

source

Garamond.rest_server — Method.

rest_server(port::Integer, io_port::Integer, search_server_ready::Condition [;ipaddr::String])

Starts a bi-directional HTTP REST server at address ipaddr::String (defaults to "0.0.0.0" i.e. all ip's) that uses the TCP port port and communicates with the search server through the TCP port io_port. The server is started once the condition search_server_ready is triggered.

source

Garamond.search — Method.

search(srcher, query [;kwargs])

Searches for query (i.e. key terms) in srcher, and returns information regarding the the documents that match best the query. The function returns an object of type SearchResult.

Arguments

srcher::Searcher is the searcher

query the query, can be either a String or Vector{String}

Keyword arguments

search_method::Symbol controls the type of matching: :exact uses exact matches while :regex consideres the needle a regular expression

max_matches::Int is the maximum number of search results to return

max_suggestions::Int is the maximum number of suggestions to return for each missing needle

source

Garamond.search — Method.

search(srchers, query [;kwargs])

Searches for query (i.e. key terms) in multiple searches and returns information regarding the documents that match best the query. The function returns the search results in the form of a Vector{SearchResult}.

Arguments

srchers::Vector{Searcher} is the searchers vector

query the query, can be either a String or Vector{String}

Keyword arguments

search_method::Symbol controls the type of matching: :exact uses exact matches while :regex consideres the needle a regular expression

max_matches::Int is the maximum number of search results to return

max_suggestions::Int is the maximum number of suggestions to return for each missing needle

custom_weights::Dict{Symbol, Float64} are custom weights for each searcher's results used in result aggregation

source

Garamond.search_server — Method.

search_server(data_config_path, io_port, search_server_ready; cache_path=nothing)

Search server for Garamond. It is a finite-state-machine that when called, creates the searchers i.e. search objects using the data_config_path and the proceeds to looping continuously in order to asynchronously handle outside requests.

After the searchers are loaded, the search server sends a notification using search_server_ready to any listening I/O servers.

source

Garamond.unix_socket_server — Method.

unix_socket_server(socket::AbstractString, io_port::Integer, start::Condition)

Starts a bi-directional unix socket server that uses a UNIX-socket socket and communicates with the search server through the TCP port io_port. The server is started once the condition start is triggered.

source

Garamond.web_socket_server — Method.

web_socket_server(port::UInt16, io_port::Integer, start::Condition [; ipaddr::String])

Starts a bi-directional web socket server that uses a WEB-socket at address ipaddr::String (defaults to "127.0.0.1") and port port and communicates with the search server through the TCP port io_port. The server is started once the condition start is triggered.

source

Garamond.DTVModel — Constant.

Constant that represents document term vector (DTV) models used in text embedding.

source

Garamond.ENVOP_REQUEST — Constant.

Request corresponding to an environment operation command.

source

Garamond.ERRORED_REQUEST — Constant.

Request corresponding to an error i.e. in parsing.

source

Garamond.EmbeddingsLibrary — Constant.

Constant that represents embeddings libraries used in text embedding.

source

Garamond.KILL_REQUEST — Constant.

Request corresponding to a kill server command.

source

Garamond.READCONFIGS_REQUEST — Constant.

Request corresponding to a searcher read configuration command.

source

Garamond.RESPONSE_TERMINATOR — Constant.

Standard response terminator. It is used in the client-server communication mark the end of sent and received messages.

source

Garamond.UNINITIALIZED_REQUEST — Constant.

Default request.

source

Garamond.BOEEmbedder — Type.

Bag-of-embeddings (BOE) structure for document embedding using word vectors.

source

Garamond.BOREPEmbedder — Type.

Bag-of-random-embedding-projections (BOREP) structure for document embedding using word vectors.

References

Wieting, Kiela ICLR 2019, "No training required: Exploring random encoders for sentence classification"

source

Garamond.BruteTreeIndex — Type.

BruteTree index type for storing text embeddings. It is a wrapper around a BruteTree NN structure and performs brute search using a distance-based similarity between vectors.

source

Garamond.CPMeanEmbedder — Type.

Concatenated-power-mean-embeddings (CPMean) structure for document embedding using word vectors.

References

Rücklé et al. 2018 "Concatenated power mean word embeddings as universal cross-lingual sentence representations"

source

Garamond.DTVEmbedder — Type.

Structure for document embedding using DTV's.

source

Garamond.DisCEmbedder — Type.

Distributed Co-occurence (DisC) structure for document embedding using word vectors.

References

Arora et al. ICLR 2018 "A compressed sensing view of unsupervised text embeddings, bag-on-n-grams and LSTMs"

source

Garamond.HNSWIndex — Type.

HNSW index type for storing text embeddings. It is a wrapper around a HierarchicalNSW (Hierarchical Navigable Small Worlds) NN graph structure and performs a very efficient search using a distance-based similarity between vectors.

References

[Y. A. Malkov, D.A. Yashunin "Efficient and robust approximate nearest

neighbor search using Hierarchical Navigable Small World graphs"] (https://arxiv.org/abs/1603.09320)

source

Garamond.InternalRequest — Type.

Request object for the internal server of the engine.

source

Garamond.KDTreeIndex — Type.

K-D Tree index type for storing text embeddings. It is a wrapper around a KDTree NN structure and performs a more efficient search using a distance-based similarity between vectors.

source

Garamond.NaiveIndex — Type.

Naive index type for storing text embeddings. It is a wrapper around a vector of embeddings and performs brute search using the cosine similarity between vectors.

source

Garamond.SIFEmbedder — Type.

Smooth inverse frequency (SIF) structure for document embedding using word vectors.

References

Arora et al. ICLR 2017, "A simple but tough-to-beat baseline for sentence embeddings"

source

Garamond.SearchEnv — Type.

Search environment object. It contains all the data, searchers
and additional structures needed by the engine to function.

source

Garamond.StringId — Type.

String ID object.

source

Base.length — Method.

length(index)

Returns the number of points indexed in index.

source

Base.parse — Method.

parse(::Type{InternalRequest}, request::AbstractString)

Parses an outside request received from a client into an InternalRequest usable by the search server.

source

Garamond.__found_missing_needles — Function.

Returns found and missing needles using an embedder

source

Garamond.aggregate! — Method.

Aggregates search results from several searchers based on their aggregation_id i.e. results from searchers with identical aggregation id's are merged together into a new search result that replaces the individual searcher ones.

source

Garamond.build_data_env — Method.

build_data_env(env::SearchEnv)

Strips searchers from env.

source

Garamond.build_logger — Function.

build_logger(logging_stream, log_level)

Builds a logger using the stream logging_streamand log_level provided.

Arguments

logging_stream::String is the output stream and can take the values:

"null" logs to /dev/null, "stdout" (default) logs to standard output, "/path/to/existing/file" logs to an existing file and "/path/to/non-existing/file" creates the log file. If no valid option is provided, the default stream is the standard output.

log_level::String is the log level can take the values "debug",

"info", "error" and defaults to "info" if no valid option is provided.

source

Garamond.build_response — Method.

build_response(dbdata, request, results, [; kwargs...])

Builds a response for an engine client using the data, request and results.

source

Garamond.build_result_from_ids — Method.

Constructs a search result from a list of data ids.

source

Garamond.build_searcher — Method.

build_searcher(dbdata, config)

Creates a Searcher from a searcher configuration config::SearchConfig.

source

Garamond.chop_to_length — Method.

Post-processes a string to fit a certain length, adding … if necessary at the end of its choped represenation.

source

Garamond.detect_language — Method.

detect_language(text [; default=DEFAULT_LANGUAGE])

Detects the language of a piece of text. Returns a language of type Languages.Language. If the text is empty of the confidence is low, return the default language.

source

Garamond.document2vec — Method.

document2vec(embedder, document [;isregex=false])

Embeds documents. The document representation is conceptually a vector of sentences, the output is always a vector of floating point numbers.

Arguments

embedder::AbstractEmbedder is the embedder

document::Vector{AbstractString} the document to be embedded, where each vector element corresponds to a sentence

Keyword arguments

isregex::Bool a false value (default) specifies that the document tokens are to be matched exactly while a true value specifies that the tokens are to be matched partially (for DTV-based document embedding only)

source

Garamond.document2vec — Method.

document2vec(embedder, document)

Word-embeddings approach to document embedding. It embeds documents using word embeddings libraries and some algorithm for combining these (depending on the type of embedder).

Arguments

embedder::WordVectorsEmbedder is the embedder

document::Vector{String} the document to be embedded, where each vector element corresponds to a sentence

source

Garamond.env_operator — Method.

env_operator(env, channels)

Saves/Loads/Updates the search environment env. Communication with the search server i.e. getting the command and its arguments and sending back a new environment is done via channels.

source

Garamond.garamond_log_formatter — Method.

garamond_log_formatter(level, _module, group, id, file, line)

Garamond -specific log message formatter. Takes a fixed set of input arguments and returns the color, prefix and suffix for the log message.

source

Garamond.noop_ranker — Method.

Noop ranker, does not rank, returns the first input argument unchanged.

source

Garamond.printable_version — Method.

printable_version()

Returns a pretty version string that includes the git commit and date.

source

Garamond.query_preparation — Method.

query_preparation(query, flags, language)

Prepares the query for search (tokenization if the case), pre-processing.

source

Garamond.read_configuration_to_json — Method.

read_configuration_to_json(env)

Returns a JSON dictionary with the full configuration of the search environment.

source

Garamond.respond — Method.

respond(env, socket, counter, channels)

Responds to search server requests received on socket using the search data from searchers. The requests are counted through the variable counter.

source

Garamond.sentences2vec — Method.

sentences2vec(embedder, document_embedding, embedded_words [;dim=0])

Returns a matrix of sentence embeddings from a vector of matrices containing individual sentence word embeddings. Used mostly for word-vectors based embedders.

Arguments

embedder::AbstractEmbedder is the embedder

document_embedding::Vector{Matrix{AbstractFloat}} are the document's word embeddings, where each element of the vector represents the embedding of a sentence (whith the matrix columns individual word embeddings)

Keyword arguments

dim::Int is the dimension of the word embeddings i.e. number of components in the word vector (default 0)

embedded_words::Vector{Vector{AbstractString}} are the words in each sentence the were embedded (their order corresponds to the order of the matrix columns in document_embedding

source

Garamond.squash — Method.

squash(m)

Function that creates a single mean vector from a matrix m and performs some normalization operations as well.

source

Garamond.squash — Method.

squash(vv, m)

Function that creates a single mean vector from a vector of vectors vv where each vector has a length m and performs some normalization operations as well.

source

Garamond.suggestion_search! — Method.

suggestion_search!(suggestions, search_tree, needles [;max_suggestions=1])

Searches in the search tree for partial matches for each of the needles.

source

Garamond.summarize — Method.

summarize(sentences [;ns=1, flags=DEFAULT_SUMMARIZATION_STRIP_FLAGS])

Build a summary of the text's sentences. The resulting summary will be a ns sentence document; each sentence is pre-procesed using the flags option.

source

Garamond.version — Method.

version()

Returns the current Garamond version using the Project.toml and git. If the Project.toml, git are not available, the version defaults to an empty string.

source

Garamond.word_embeddings — Method.

word_embeddings(word_vectors, document_tokens [;kwargs])

Returns a matrix corresponding to the word embeddings of document_tokens as well as the indices of missing i.e. not-embedded tokens.

Arguments

word_vectors::EmbeddingsLibrary wordvectors object; can be a Word2Vec.WordVectors, Glowe.WordVectors or ConceptnetNumberbatch.ConceptNet

document_tokens::Vector{String} the words to be embedded, where each vector element corresponds to a word

Keyword arguments

keep_size::Bool a false value discards vectors for words not found while a true value (default) places a zero vector in the embeddings matrix

print_matched_words::Bool if true, the words that were and that were not embedded are printed (default false)

kwargs... the rest of the keyword arguments are ConceptNet specific and can be found by inspecting the help of ConceptnetNumberbatch.embed_document

source

HNSW.knn_search — Method.

knn_search(index, point, k, keep)

Searches for the k nearest neighbors of point in data contained in the index. The index may vary from a simple wrapper inside a matrix to more complex structures such as k-d trees, etc. Only neighbors present in keep are returned.

source