API Reference

Garamond.ERRORED_REQUEST

Garamond.BruteTreeEmbeddingModel

Garamond.HNSWEmbeddingModel

Garamond.KDTreeEmbeddingModel

Garamond.NaiveEmbeddingModel

ConceptnetNumberbatch.embed_document

ConceptnetNumberbatch.embed_document

Garamond.build_corpus

Garamond.build_logger

Garamond.build_searcher

Garamond.construct_response

Garamond.garamond_log_formatter

Garamond.get_parsing_function

Garamond.ioserver

Garamond.load_search_configs

Garamond.prepare_query

Garamond.printable_version

Garamond.recursive_glob

Garamond.search

Garamond.search

Garamond.search

Garamond.search_heuristically!

Garamond.search_server

Garamond.squash

Garamond.squash

Garamond.summarize

Garamond.unix_socket_server

Garamond.updater

Garamond.version

Garamond.web_socket_server

Garamond.search — Method.

search(model, point, k)

Searches for the k nearest neighbors of point in data contained in the model. The model may vary from a simple wrapper inside a matrix to more complex structures such as k-d trees, etc.

source

Garamond.search — Method.

search(termcnt, needles, method)

Search function for searching using the term imporatances associated to a corpus.

source

Garamond.search — Method.

search(srcher, query [;kwargs])

Searches for query (i.e. key terms) in multiple corpora and returns information regarding the documents that match best the query. The function returns the search results in the form of a Vector{SearchResult}.

Arguments

srcher::AbstractVector{AbstractSearcher} is the corpora searcher

query the query

Keyword arguments

search_type::Symbol is the type of the search; can be :metadata, :data or :all; the options specify that the query can be found in the metadata of the documents of the corpus, the document content or both respectively

search_method::Symbol controls the type of matching: :exact searches for the very same string while :regex searches for a string in the corpus that includes the needle

max_matches::Int is the maximum number of search results to return from each corpus

max_corpus_suggestions::Int is the maximum number of suggestions to return for each missing needle from the search in a corpus

source

Garamond.ERRORED_REQUEST — Constant.

deconstruct_request(request)

Function that deconstructs a Garamond request received from a client into individual search engine operations and search parameters.

source

Garamond.BruteTreeEmbeddingModel — Type.

BruteTree model type for storing text embeddings. It is a wrapper around a BruteTree NN structure and performs brute search using a distance-based similarity between vectors.

source

Garamond.HNSWEmbeddingModel — Type.

HNSW model type for storing text embeddings. It is a wrapper around a HierarchicalNSW (Hierarchical Navigable Small Worlds [1]) NN graph structure and performs a very efficient search using a distance-based similarity between vectors. [1] Yu. A. Malkov, D.A. Yashunin "Efficient and robust approximate nearest neighbor search using Hierarchical Navigable Small World graphs" (https://arxiv.org/abs/1603.09320)

source

Garamond.KDTreeEmbeddingModel — Type.

K-D Tree model type for storing text embeddings. It is a wrapper around a KDTree NN structure and performs a more efficient search using a distance-based similarity between vectors.

source

Garamond.NaiveEmbeddingModel — Type.

Naive model type for storing text embeddings. It is a wrapper around a matrix of embeddings and performs brute search using the cosine similarity between vectors.

source

ConceptnetNumberbatch.embed_document — Method.

Function that embeds a document i.e. returns an embedding matrix, columns are word embeddings, using the Word2Vec or Glowe WordVectors object. The function has an identical signature as the one from the ConceptnetNumberbatch package.

source

ConceptnetNumberbatch.embed_document — Method.

embed_document(embeddings_library, lexicon, document[; embedding_method])

Function to get from multiple sentencea to a document embedding. The embedding_method option controls how multiple sentence embeddings are combined into a single document embedding. Avalilable options for embedding_method: :bow - calculates document embedding as the mean of the sentence embeddings :sif - smooth-inverse-frequency subtracts paragraph/phrase vector from each sentence embedding

source

Garamond.build_corpus — Method.

build_corpus(documents, doctype, metadata_vector)

Builds a corpus of documents of type doctype using the data in documents and metadata from metadata_vector.

Note: No preprocessing is performed at this step, it is assumed that the data has already been preprocessed and is ready to be searched in.

source

Garamond.build_logger — Function.

build_logger(logging_stream, log_level)

Builds a logger using the stream and loglevel provided. These should be coming from parsing the input arguments. The loggingstream can take the values: • "null": logs to /dev/null • "stdout": logs to standard output • "/path/to/existing/file": logs to an existing file • "/path/to/non-existing/file": creates the log file If no valid option is provided, the default stream is the standard output. The log level can take the values: "debug", "info", "error" and defaults to "info" if no valid option is provided.

source

Garamond.build_searcher — Method.

build_searcher(sconf)

Creates a Searcher from a SearchConfig.

source

Garamond.construct_response — Method.

construct_response(srchers, results, what [; kwargs...])

Function that constructs a response for a Garamond client using the search results, data from srchers and specifier what.

source

Garamond.garamond_log_formatter — Method.

garamond_log_formatter(level, _module, group, id, file, line)

Garamond -specific log message formatter. Takes a fixed set of input arguments and returns the color, prefix and suffix for the log message.

source

Garamond.get_parsing_function — Method.

get_parsing_function(args...)

Function that generates a parsing function from its input arguments and returns it.

Arguments

parser::Symbol is the name of the parser

parser_config::Union{Nothing, Dict} can contain optional configuration data for the parser (for delimited parsers)

header::Bool whether the file has a header or not (for delimited files only)

delimiter::String the delimiting character (for delimited files only)

globbing_pattern::String globbing pattern for gathering file lists from directories (for directory parsers only)

build_summary::Bool whether to use a summary instead of the full document (for directory parsers only)

summary_ns::Int how many sentences to use in the summary (for directory parsers only)

summarization_strip_flags::UInt32 flags used to strip text before summarization (for directory parsers only)

show_progress::Bool whether to show the progress when loading files

Note: parser must be in the keys of the PARSER_CONFIGS constant. The name of the data parsing function is created as: :__parser_<parser> so, the function name :__parser_delimited_format_1 corresponds to the parser :delimited_format_1. The function must be defined apriori.

source

Garamond.ioserver — Method.

ioserver(socket_or_port::Union{UInt16, AbstractString}, channel::Channel{String})

Wrapper for the UNIX- or WEB- socket servers.

source

Garamond.load_search_configs — Method.

load_search_configs(filename)

Function that creates search configurations from a data configuration file specified by filename. It returns a Vector{SearchConfig} that is used to build the Searcher objects with which search is performed.

source

Garamond.prepare_query — Method.

prepare_query(query, flags)

Prepares the query for search (tokenization if the case), pre-processing.

source

Garamond.printable_version — Method.

printable_version()

Returns a pretty version string that includes the git commit and date.

source

Garamond.recursive_glob — Function.

recursive_glob(pattern, path)

Globs recursively all the files matching the pattern, at the given path.

source

Garamond.search_heuristically! — Method.

search_heuristically!(suggestions, search_tree, needles [;max_suggestions=1])

Searches in the search tree for partial matches of the needles.

source

Garamond.search_server — Method.

search_server(data_config_paths, socket, ws_port)

Main server function of Garamond. It is a finite-state-machine that when called, creates the searchers i.e. search objects using the data_config_paths and the proceeds to looping continuously in order to: • update the searchers regularly; • receive requests from clients using the unix-socket socket or/and a web-socket at port ws_port; • call search and route responses back to the clients through their corresponding sockets

Both searcher update and I/O communication are performed asynchronously.

source

Garamond.squash — Method.

squash(m)

Function that creates a single mean vector from a matrix m and performs some normalization operations as well.

source

Garamond.squash — Method.

squash(vv, m)

Function that creates a single mean vector from a vector of vectors vv where each vector has a length m and performs some normalization operations as well.

source

Garamond.summarize — Method.

summarize(sentences [;ns=1, flags=DEFAULT_SUMMARIZATION_FLAGS])

Build a summary of the text's sentences. The resulting summary will be a ns sentence document; each sentence is pre-procesed using the flags option.

source

Garamond.unix_socket_server — Method.

unix_socket_server(socket::AbstractString, channel::Channel{String})

Starts a bi-directional unix socket server that uses a UNIX-socket socket and communicates with the search server through a channel channel.

source

Garamond.updater — Method.

updater(searchers, channel, update_interval)

Function that regularly updates the searchers at each update_interval seconds, and puts the updates on the channel to be sent to the search server.

source

Garamond.version — Method.

version()

Returns the current Garamond version using the Project.toml and git. If the Project.toml, git are not available, the version defaults to an empty string.

source

Garamond.web_socket_server — Method.

web_socket_server(port::UInt16, channel::Channel{String})

Starts a bi-directional web socket server that uses a WEB-socket at port port and communicates with the search server through a channel channel.

source